Fixed a bug in the construct function in the Sprite class. #Adam Labranche <adam.labranche@anhero.net> Fixed some bugs about the vertices construction in the Sprite class. There was a problem with the variable arguments function for adding vertices to a VerticesGroup. Still more bugs need to be fixed.

+++ RedBox/other/VerticesGroup.cpp
@@ -60,8 +60,8 @@ void VerticesGroup::addVertices(unsigned int nbVertices, ...) {
 	unsigned int initialSize = verticesData.size();
 	verticesData.resize(verticesData.size() + nbVertices * 2);
 	for(unsigned int i = 0; i < nbVertices; i++) {
-		verticesData[i * 2 + initialSize] = va_arg(verticesCoords, float);
-		verticesData[i * 2 + 1 + initialSize] = va_arg(verticesCoords, float);
+		verticesData[i * 2 + initialSize] = static_cast<float>(va_arg(verticesCoords, double));
+		verticesData[i * 2 + 1 + initialSize] = static_cast<float>(va_arg(verticesCoords, double));
 	}
 	va_end(verticesCoords);
